fix: vote에 축약된 카드 분리(#30)
